feat(#534): disassemble annotation default value
